+++ ricaun.RevitTest.TestAdapter/Extensions/Json/IJsonService.cs
@@ -1,0 +1,8 @@
+﻿namespace ricaun.RevitTest.TestAdapter.Extensions.Json
+{
+    internal interface IJsonService
+    {
+        T Deserialize<T>(string value);
+        string Serialize(object value);
+    }
+}