+++ ricaun.RevitTest.TestAdapter/TestAdapter.cs
@@ -30,7 +30,11 @@ namespace ricaun.RevitTest.TestAdapter
             AdapterSettings.Create(discoveryContext);
             AdapterLogger.Create(messageLogger, AdapterSettings.Settings.NUnit.Verbosity);
 
+            var targetFrameworkAttribute = this.GetType().Assembly
+               .GetCustomAttribute(typeof(System.Runtime.Versioning.TargetFrameworkAttribute)) as System.Runtime.Versioning.TargetFrameworkAttribute;
+
             AdapterLogger.Logger.Info($"TestAdapter: {this.AdapterVersion}", 0);
+            AdapterLogger.Logger.Info($"TargetFramework: {targetFrameworkAttribute?.FrameworkName}");
             AdapterLogger.Logger.Info($"AdapterSettings: {AdapterSettings.Settings}");
 
             if (AdapterSettings.Instance.RunSettings is null)
@@ -41,6 +45,7 @@ namespace ricaun.RevitTest.TestAdapter
 #if DEBUG
             AdapterLogger.Logger.DebugOnlyLocal("-DEBUG-");
             AdapterLogger.Logger.DebugOnlyLocal($"\tTestAdapter: {this.AdapterVersion}");
+            AdapterLogger.Logger.DebugOnlyLocal($"\tTargetFramework: {targetFrameworkAttribute?.FrameworkName}");
             AdapterLogger.Logger.DebugOnlyLocal($"\tAdapterSettings: {AdapterSettings.Settings}");
 
             var collection = Metadatas.XmlUtils.ParseKeyValues(discoveryContext.RunSettings.SettingsXml);